+++ package.json
@@ -15,6 +15,7 @@
     "@emotion/styled": "^11.14.1",
     "@mui/icons-material": "^7.2.0",
     "@mui/material": "^7.2.0",
+    "axios": "^1.11.0",
     "next": "15.4.1",
     "react": "19.1.0",
     "react-dom": "19.1.0",
