+++ src/apis/auth.ts
@@ -1,0 +1,24 @@
+import axios from "axios";
+
+export interface LoginPayload {
+  id: string;
+  password: string;
+}
+
+export const login = async ({ id, password }: LoginPayload) => {
+  // const params = new URLSearchParams();
+  // params.append("id", id);
+  // params.append("password", password);
+
+  const params = `id=${id}&password=${password}`;
+
+  console.log(params);
+  const response = await axios.post(`${process.env.NEXT_PUBLIC_API_URL}/auth/signin`, params, {
+    headers: {
+      "Content-Type": "application/x-www-form-urlencoded",
+    },
+    withCredentials: true, // 쿠키로 세션 유지
+  });
+
+  return response.data;
+};
