+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from "./panel/LoginForm";
+import LoginForm from "./login/LoginForm";
 
 export default function LoginPage() {
   return (
